phase_02/backend/pyproject.toml
[project]
name = "plannoir-backend"
version = "0.1.0"
description = "Backend API for Plannoir - Premium AI-Native Todo Full-Stack Web Application"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "sqlmodel>=0.0.8",
    "pyjwt>=2.8.0",
    "psycopg2-binary>=2.9.7",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "psycopg2>=2.9.11",
    "pytest>=9.0.2",
    "alembic>=1.18.0",
    "passlib>=1.7.4",
    "bcrypt>=5.0.0",
    "httpx>=0.28.1",
    "email-validator>=2.3.0",
    "requests>=2.32.5",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
